fix get maintenance

+++ BeerMachineApi/Services/StatusModels/MaintenanceStatusModel.cs
@@ -9,14 +9,14 @@ public class MaintenanceStatusModel
         MaintenanceState = false;
     }
 
-    public float MaintenanceCount { get; set; }
-    public int MaintenanceTrigger { get; set; }
+    public System.UInt16 MaintenanceCount { get; set; }
+    public System.UInt16 MaintenanceTrigger { get; set; }
     public bool MaintenanceState { get; set; }
 
     public void UpdateModel(Opc.UaFx.Client.OpcClient session)
     {
-        MaintenanceCount = (int)session.ReadNode(NodeIds.MaintenanceCount).Value;       // current amount 
-        MaintenanceTrigger = (int)session.ReadNode(NodeIds.MaintenanceTrigger).Value;   // max maintenance amount 
+        MaintenanceCount = (System.UInt16)session.ReadNode(NodeIds.MaintenanceCount).Value;       // current amount 
+        MaintenanceTrigger = (System.UInt16)session.ReadNode(NodeIds.MaintenanceTrigger).Value;   // max maintenance amount 
 
         if ((byte)session.ReadNode(NodeIds.MaintenanceState).Value == 20 || MaintenanceCount >= MaintenanceTrigger)
             MaintenanceState = true;   // during maintenance
